Multiple qubits in Results
