+++ src/qibocal/protocols/characterization/__init__.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+from .resonator_spectroscopy import resonator_spectroscopy
+
+
+class Operation(Enum):
+    resonator_spectroscopy = resonator_spectroscopy
